refactor(activity): extract AddToFavorites button into separate component
